backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.2.2

# Database
motor==3.3.1
pymongo==4.5.0

# Validation
pydantic==2.12.5
email-validator==2.3.0

# Auth / security
passlib==1.7.4
python-jose==3.5.0

# Requests (Mailgun)
requests==2.32.5

# Utils
python-multipart==0.0.22